fix(API docs): Add note about CLI release management to permissions page (#860)
A user trying to associate commits using the CLI got a 403 and came to the API permissions page to see which scopes were necessary for release endpoints. Though technically nothing on that page was incorrect (the CLI hits an org endpoint when executing `set-commits`, and its scope _is_ correctly listed), someone in the user's position would have no way to know to look anywhere besides the the release permissions.

This adds a note explaining the additional required scope.