ref: Update Jekyll
This locks Jekyll and its deps at a stable version, removes a bunch of unused content, and tweaks a couple of things for compatibility.

While the many hours that have passed in achieving this have caused me to forget the original purpose of this PR, I assure you it's in service of removing Jekyll in favor of Gatsby.