feat: Template variables in markdown (#2015)
This primarily adds support for template variables (or otherwise javascript functions) executable within any markdown content. An example is found in the ``README.md``. Generally speaking, its the same as Django or Liquid: ``{{ scoped_variable }}``.

Using the system it implements the PackageRegistry and adds it to the scope as ``packages``, thus making them callable with e.g. ``packages.version()``.

Additionally this makes a few small changes to other various Gatsby configurations to improve performance.