Create Open API docs v1 (#2137)
Objective
We want to create API docs from an OpenAPI document. First we created a gatsby-plugin to convert openapi-derefed.json from getsentry/sentry-api-schema into graphql nodes. Next, we created a new docs page under the route /development-api/*. This is temporary and we will be taking over the /api/* routes. Next we created a robots.txt file and disallowed crawling /development-api/*.

Note: We resolve a specific commit of getsentry/sentry-api-schema into graphql nodes. Currently, this SHA needs to be updated manually.