feat: Remove unused code related to nginx (#2377)
## Objective
We don't use nginx for redirects. Redirects are handled through: https://docs.sentry.io/contributing/pages/redirects/

So in this PR, unused code related to nginx and its' redirects should be removed.

Consensus is that the Dockerfile is not necessary, so it will be also removed in this PR.

## Test Plan
Check the Vercel preview build to make sure redirects did not break.